+++ jgclark.Dashboard/src/react/components/Header/Header.jsx
@@ -2,7 +2,7 @@
 //--------------------------------------------------------------------------
 // Dashboard React component to show the Header at the top of the Dashboard window.
 // Called by Dashboard component.
-// Last updated 2024-08-05 for v2.1.0.a4 by @dbw
+// Last updated 2024-08-13 for v2.1.0.a7 by @dbw
 //--------------------------------------------------------------------------
 
 //--------------------------------------------------------------------------
@@ -106,6 +106,7 @@ const Header = ({ lastFullRefresh }: Props): React$Node => {
   const isDesktop = pluginData.platform === "macOS"
   const updatedText = "Updated"
   const timeAgoText = isDesktop ? timeAgo : timeAgo.replace(" mins", "m").replace(" min", "m")
+  const isNarrowWidth = window.innerWidth <= 600
 
   return (
     <div className="header">
@@ -120,7 +121,7 @@ const Header = ({ lastFullRefresh }: Props): React$Node => {
             className="HAButton hardRefreshButton"
           >
             <i className={"fa-regular fa-arrows-retweet"}></i>
-            <span className="pad-left">{isDesktop ? "Hard Refresh" : " HR "}</span>
+            <span className="pad-left">{isNarrowWidth ? "HR" : "Hard Refresh"}</span>
           </button>
         )}
       </div>