+++ jgclark.Dashboard/CHANGELOG.md
@@ -1,7 +1,14 @@
 # What's changed in 🎛 Dashboard plugin?
 For more details see the [plugin's documentation](https://github.com/NotePlan/plugins/tree/main/jgclark.Dashboard/).
+## [Perspectives.a7] @jgc, 2024-08-13 WIP
+- removed some of the circular dependencies
+- updated names of some settings, particularly `ignoreFolders`
+- added updateCurrentPerspectiveDef(), and "/Update current Perspective" command for testing. FIXME: is failing to read dashboardSettings properly.
+- added savePerspectiveSettings() to make one place to save changes (at least on the back end). FIXME: need to get this to affect front end too.
+- turned off underlining on the 'take a break' message lines
+- clicking on 'there are X items hidden' message lines now should turn off filtering. FIXME: but it doesn't persist or update the window.
 
-## [Perspectives.a6] @jgc,  2024-08-10 WIP for @dbw
+## [Perspectives.a6] @jgc, 2024-08-10 WIP for @dbw
 - fixed perspectiveSetting initialisation in WebView (I think) and how to persist it in Dashboard::useEffects (I think)
 - BUT FIXME: changing PerspectiveSelector logs the right newValue, but it then fails to get updated anywhere else
 - dealt with FIXME in dataGeneration about calling `getProjectSectionData`
@@ -10,7 +17,7 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 - added subtle hover effect to UI buttons, using the clever new CSS 'hsl' operator
 - changed properly to "Dashboard.css" (capital D) to match the Component name
 
-## [Perspectives.a5] @jgc,  2024-08-09 WIP for @dbw
+## [Perspectives.a5] @jgc, 2024-08-09 WIP for @dbw
 - made a separate perspectiveSettings data structure, than now includes all settings, not just the 5 from the proof of concept. @dbw: something is wrong with the initialisation of perspectives.
 - tighten up removal of priority indicators, to only happen at the start of a line's content
 - now won't display buttons in the Section header if there are no items to work on. (However, the 'add' buttons in the calendar sections are still shown.)